Use make_msg

+++ src/rpc/full_node_rpc_api.py
@@ -106,7 +106,7 @@ class FullNodeRpcApi:
         if (
             full_peak is None
             or full_peak.foliage_block is None
-            or full_peak.foliage_block.timestamp < now - 60 * 10
+            or full_peak.foliage_block.timestamp < int(now - 60 * 10)
             or sync_mode
         ):
             synced = False
